Replace all old tests with the new ones
