+++ test.sh
@@ -2,5 +2,5 @@
 
 docker pull ubuntu:latest
 ./node_modules/standard/bin/cmd.js src
-./node_modules/mocha/bin/mocha --require babel-core/register -R spec
+./node_modules/.bin/ava --timeout 30000 test
 exit $?
