+++ test/container.js
@@ -1,12 +1,18 @@
-import chai from 'chai'
-import * as test_utils from './utils'
+import test from 'ava'
+import fs from 'fs'
 import Container from '../lib/container'
 import Image from '../lib/image'
 import {default as MemoryStream} from 'memorystream'
+import { Docker } from '../lib/docker'
 
-const should = chai.should()
+const socket = process.env.DOCKER_SOCKET || '/var/run/docker.sock'
+const isSocket = fs.existsSync(socket) ? fs.statSync(socket).isSocket() : false
+const docker = isSocket
+  ? new Docker()
+  : new Docker({ socketPath: socket })
 
-const docker = test_utils.init()
+const createContainer = (name, extra) =>
+  docker.container.create(Object.assign({ Image: 'ubuntu', name: containerNames.get(name) }, extra))
 const containerNames = new Map([
   [ 'create', 'docker_api_test_create' ],
   [ 'inspect', 'docker_api_test_inspect' ],
@@ -22,6 +28,7 @@ const containerNames = new Map([
   [ 'kill', 'docker_api_test_kill' ],
   [ 'update', 'docker_api_test_update' ],
   [ 'rename', 'docker_api_test_rename' ],
+  [ 'rename_prev', 'docker_api_test_rename_prev' ],
   [ 'pause', 'docker_api_test_pause' ],
   [ 'attach', 'docker_api_test_attach' ],
   [ 'commit', 'docker_api_test_commit' ],
@@ -32,484 +39,240 @@ const containerNames = new Map([
   [ 'info_archive', 'docker_api_test_info_archive' ]
 ])
 
-describe('#container', function () {
-  describe('#list', function () {
-    it('should receive an array of containers', function () {
-      this.timeout(30000)
-      return docker.container.list()
-        .then((containers) => {
-          containers.should.be.an('array')
-        })
-    })
-  })
+test('list', async t => {
+  const containers = await docker.container.list()
+  t.is(containers.constructor, Array)
+})
 
-  describe('#create', function () {
-    it('should create a container', function () {
-      this.timeout(30000)
-      return docker.container.create({ Image: 'ubuntu', name: containerNames.get('create') })
-        .then((container) => {
-          container.should.be.instanceof(Container)
-        })
-    })
-  })
-  
-  describe('#inspect', function () {
-    it('should inspect a container', function () {
-      this.timeout(30000)
-      return docker.container.create({ Image: 'ubuntu', name: containerNames.get('inspect') })
-        .then((container) => {
-          return container.status()
-        })
-        .then((container) => {
-          container.should.be.instanceof(Container)
-        })
-    })
-  })
-  
-  describe('#top', function () {
-    it('should get top processes', function () {
-      this.timeout(30000)
-      return docker.container.create({ 
-        Image: 'ubuntu', 
-        Cmd: [ '/bin/bash', '-c', 'tail -f /var/log/dmesg' ], 
-        name: containerNames.get('top') 
-      })
-        .then((container) => {
-          return container.start()
-        })
-        .then((container) => {
-          return container.top()
-        })
-        .then((processes) => {
-          processes.Processes.should.be.an('array')
-        })
-    })
-  })
-  
-  describe('#logs', function () {
-    it('should get container log', function () {
-      this.timeout(30000)
-      return docker.container.create({ 
-        Image: 'ubuntu', 
-        Cmd: [ '/bin/bash', '-c', 'tail -f /var/log/dmesg' ], 
-        name: containerNames.get('logs') 
-      })
-        .then((container) => {
-          return container.start()
-        })
-        .then((container) => {
-          return container.logs({ stdout: 1, follow: true })
-        })
-        .then((logs) => {
-          logs.pipe.should.exist
-        })
-    })
-  })
-  
-  describe('#changes', function () {
-    it('should get container log', function () {
-      this.timeout(30000)
-      return docker.container.create({ 
-        Image: 'ubuntu', 
-        Cmd: [ '/bin/bash', '-c', 'echo "xfoo" > foo.txt' ], 
-        name: containerNames.get('changes') 
-      })
-        .then((container) => {
-          return container.start()
-        })
-        .then((container) => {
-          return container.changes()
-        })
-        .then((changes) => {
-          changes.should.be.an('array')
-        })
-    })
-  })
-  
-  describe('#export', function () {
-    it('should export a container', function () {
-      this.timeout(30000)
-      return docker.container.create({ Image: 'ubuntu', name: containerNames.get('export') })
-        .then((container) => {
-          return container.export({ stream: false })
-        })
-        .then((container) => {
-          container.should.be.ok
-        })
-    })
-  })
-  
-  describe('#stats', function () {
-    it('should get stats of a container', function () {
-      this.timeout(70000)
-      return docker.container.create({ 
-        Image: 'ubuntu', 
-        Cmd: [ '/bin/bash', '-c', 'tail -f /var/log/dmesg' ], 
-        name: containerNames.get('stats') 
-      })
-        .then((container) => {
-          return container.start()
-        })
-        .then((container) => {
-          return container.stats()
-        })
-        .then((stats) => {
-          stats.pipe.should.be.ok
-        })
-    })
-  })
+test('should create a container', async t => {
+  const container = await createContainer('create')
+  t.is(container.constructor, Container)
+})
 
-  describe('#resize', function () {
-    it('should resize tty a container', function () {
-      this.timeout(70000)
-      return docker.container.create({ 
-        Image: 'ubuntu', 
-        Cmd: [ '/bin/bash', '-c', 'tail -f /var/log/dmesg' ], 
-        name: containerNames.get('resize') 
-      })
-        .then((container) => {
-          return container.start()
-        })
-        .then((container) => {
-          return container.stats({
-            h: process.stdout.rows,
-            w: process.stdout.columns
-          })
-        })
-        .then((res) => {
-          res.pipe.should.be.ok
-        })
-    })
+test('inspect', async t => {
+  const container = await createContainer('inspect')
+  const containerStatus = await container.status()
+  t.is(containerStatus.constructor, Container)
+})
+
+test('top', async t => {
+  const container = await createContainer('top', {
+    Cmd: [ '/bin/bash', '-c', 'tail -f /var/log/dmesg' ], 
   })
+  await container.start()
+  const processes = await container.top()
+  t.is(processes.Processes.constructor, Array)
+})
 
-  describe('#start', function () {
-    it('should start a container', function () {
-      this.timeout(70000)
-      return docker.container.create({ 
-        Image: 'ubuntu', 
-        Cmd: [ '/bin/bash', '-c', 'tail -f /var/log/dmesg' ], 
-        name: containerNames.get('start') 
-      })
-        .then((container) => {
-          return container.start()
-        })
-        .then((container) => {
-          container.should.be.instanceof(Container)
-        })
-    })
+test('log', async t => {
+  const container = await createContainer('logs', {
+    Cmd: [ '/bin/bash', '-c', 'tail -f /var/log/dmesg' ], 
   })
+  await container.start()
+  const logs = await container.logs({ stdout: 1, follow: true })
+  t.truthy(logs.pipe)
+})
 
-  describe('#stop', function () {
-    it('should stop a container', function () {
-      this.timeout(70000)
-      return docker.container.create({ 
-        Image: 'ubuntu', 
-        Cmd: [ '/bin/bash', '-c', 'tail -f /var/log/dmesg' ], 
-        name: containerNames.get('stop') 
-      })
-        .then((container) => {
-          return container.start()
-        })
-        .then((container) => {
-          return container.stop()
-        })
-        .then((container) => {
-          return container.status()
-        })
-        .then((container) => {
-          container.State.Status.should.equal('exited')
-        })
-    })
+test('changes', async t => {
+  const container = await createContainer('changes', {
+    Cmd: [ '/bin/bash', '-c', 'echo "xfoo" > foo.txt' ], 
   })
+  await container.start()
+  const changes = await container.changes()
+  t.is(changes.constructor, Array)
+})
 
-  describe('#restart', function () {
-    it('should restart a container', function () {
-      this.timeout(70000)
-      return docker.container.create({ 
-        Image: 'ubuntu', 
-        Cmd: [ '/bin/bash', '-c', 'tail -f /var/log/dmesg' ], 
-        name: containerNames.get('restart')
-      })
-        .then((container) => {
-          return container.start()
-        })
-        .then((container) => {
-          return container.restart()
-        })
-        .then((container) => {
-          return container.status()
-        })
-        .then((container) => {
-          container.State.Status.should.equal('running')
-        })
-    })
+test('export', async t => {
+  const container = await createContainer('export')
+  const result = container.export({ stream: false })
+  t.truthy(result)
+})
+
+test('stats', async t => {
+  const container = await createContainer('stats', {
+    Cmd: [ '/bin/bash', '-c', 'tail -f /var/log/dmesg' ], 
   })
+  await container.start()
+  const stats = await container.stats()
+  t.truthy(stats.pipe)
+})
 
-  describe('#kill', function () {
-    it('should kill a container', function () {
-      this.timeout(70000)
-      return docker.container.create({ 
-        Image: 'ubuntu', 
-        Cmd: [ '/bin/bash', '-c', 'tail -f /var/log/dmesg' ], 
-        name: containerNames.get('kill') 
-      })
-        .then((container) => {
-          return container.start()
-        })
-        .then((container) => {
-          return container.kill()
-        })
-        .then((container) => {
-          return container.status()
-        })
-        .then((container) => {
-          container.State.Status.should.equal('exited')
-        })
-    })
+test('resize', async t => {
+  const container = await createContainer('resize', {
+    Cmd: [ '/bin/bash', '-c', 'tail -f /var/log/dmesg' ], 
   })
-
-  describe('#update', function () {
-    it('should update a container', function () {
-      this.timeout(70000)
-      return docker.container.create({ 
-        Image: 'ubuntu', 
-        Cmd: [ '/bin/bash', '-c', 'tail -f /var/log/dmesg' ], 
-        name: containerNames.get('update') 
-      })
-        .then((container) => {
-          return container.update({ 'CpuShares': 512 })
-        })
-        .then((container) => {
-          return container.status()
-        })
-        .then((container) => {
-          container.HostConfig.CpuShares.should.equal(512)
-        })
-    })
+  await container.start()
+  const result = await container.stats({
+    h: process.stdout.rows,
+    w: process.stdout.columns
   })
+  t.truthy(result.pipe)
+})
 
-  describe('#rename', function () {
-    it('should rename a container', function () {
-      this.timeout(70000)
-      return docker.container.create({ 
-        Image: 'ubuntu', 
-        Cmd: [ '/bin/bash', '-c', 'tail -f /var/log/dmesg' ], 
-        name: containerNames.get('rename') + '_prev' 
-      })
-        .then((container) => {
-          return container.rename({ 'name': containerNames.get('rename') })
-        })
-    })
+test('start', async t => {
+  const container = await createContainer('start', {
+    Cmd: [ '/bin/bash', '-c', 'tail -f /var/log/dmesg' ], 
   })
+  const result = await container.start()
+  t.is(result.constructor, Container)
+})
 
-  describe('#pause', function () {
-    it('should pause a container', function () {
-      this.timeout(70000)
-      return docker.container.create({ 
-        Image: 'ubuntu', 
-        Cmd: [ '/bin/bash', '-c', 'tail -f /var/log/dmesg' ], 
-        name: containerNames.get('pause') 
-      })
-        .then((container) => {
-          return container.start()
-        })
-        .then((container) => {
-          return container.pause()
-        })
-        .then((container) => {
-          return container.status()
-        })
-        .then((container) => {
-          container.State.Status.should.equal('paused')
-        })
-    })
+test('stop', async t => {
+  const container = await createContainer('stop', {
+    Cmd: [ '/bin/bash', '-c', 'tail -f /var/log/dmesg' ], 
   })
+  await container.start()
+  await container.stop()
+  const containerStatus = await container.status()
+  t.is(containerStatus.State.Status, 'exited')
+})
 
-  describe('#unpause', function () {
-    it('should unpause a container', function () {
-      this.timeout(70000)
-      return docker.container.status(containerNames.get('pause'))
-        .then((container) => {
-          return container.unpause()
-        })
-        .then((container) => {
-          return container.status()
-        })
-        .then((container) => {
-          container.State.Status.should.equal('running')
-        })
-    })
+test('restart', async t => {
+  const container = await createContainer('restart', {
+    Cmd: [ '/bin/bash', '-c', 'tail -f /var/log/dmesg' ], 
   })
+  await container.start()
+  await container.restart()
+  const containerStatus = await container.status()
+  t.is(containerStatus.State.Status, 'running')
+})
 
-  describe('#commit', function () {
-    it('should commit changes in a container', function () {
-      this.timeout(70000)
-      return docker.container.create({ 
-        Image: 'ubuntu', 
-        Cmd: [ '/bin/bash', '-c', 'tail -f /var/log/dmesg' ], 
-        name: containerNames.get('commit') 
-      })
-        .then((container) => {
-          return container.start()
-        })
-        .then((container) => {
-          return container.commit({ comment: 'commit test' })
-        })
-        .then((image) => {
-          image.should.be.instanceof(Image)
-        })
-    })
+test('kill', async t => {
+  const container = await createContainer('kill', {
+    Cmd: [ '/bin/bash', '-c', 'tail -f /var/log/dmesg' ], 
   })
+  await container.start()
+  await container.kill()
+  const containerStatus = await container.status()
+  t.is(containerStatus.State.Status, 'exited')
+})
 
-  describe('#exec', function () {
-    it('should run exec on a container', function () {
-      this.timeout(70000)
-      return docker.container.create({ 
-        Image: 'ubuntu', 
-        Cmd: [ '/bin/bash', '-c', 'tail -f /var/log/dmesg' ], 
-        name: containerNames.get('exec')
-      })
-        .then((container) => {
-          return container.start()
-        })
-        .then((container) => {
-          return container.exec.create({
-            Cmd: [ "top" ]
-          })
-        })
-        .then((exec) => {
-          return exec.start()
-        })
-        .then((stream) => {
-          stream.pipe.should.be.ok
-        })
-    })
-
-    it('should check status of exec on a container', function () {
-      this.timeout(70000)
-      let Exec
-      return docker.container.create({ 
-        Image: 'ubuntu', 
-        Cmd: [ '/bin/bash', '-c', 'tail -f /var/log/dmesg' ], 
-        name: containerNames.get('inspect_exec')
-      })
-        .then((container) => {
-          return container.start()
-        })
-        .then((container) => {
-          return container.exec.create({
-            Cmd: [ "top" ]
-          })
-        })
-        .then((exec) => {
-          Exec = exec.constructor
-          return exec.status()
-        })
-        .then((exec) => {
-          exec.should.be.instanceof(Exec)
-        })
-    })
+test('update', async t => {
+  const container = await createContainer('update', {
+    Cmd: [ '/bin/bash', '-c', 'tail -f /var/log/dmesg' ], 
   })
+  await container.update({ 'CpuShares': 512 })
+  const containerStatus = await container.status()
+  t.is(containerStatus.HostConfig.CpuShares, 512)
+})
 
-  describe('#attach', function () {
-    it('should attach and wait for a container', function () {
-      this.timeout(70000)
-      return docker.container.create({
-        'AttachStdin': false,
-        'AttachStdout': true,
-        'AttachStderr': true,
-        'Tty': false,
-        'OpenStdin': false,
-        'StdinOnce': false,
-        'Env': null,
-        'Cmd': ['/bin/bash', '-c', 'uptime'],
-        'Dns': ['8.8.8.8', '8.8.4.4'],
-        'Image': 'ubuntu',
-        name: containerNames.get('attach')
-      })
-        .then((container) => {
-          return container.attach({
-            stream: true,
-            stdout: true,
-            stderr: true
-          })
-        })
-        .then((result) => {
-          const [ stream, container ] = result
-          stream.should.be.ok
-
-          return container.start()
-        })
-        .then((container) => {
-          return container.wait()
-        })
-        .then((code) => {
-          code.StatusCode.should.be.a('number')
-        })
-    })
+test('rename', async t => {
+  const container = await createContainer('rename_prev', {
+    Cmd: [ '/bin/bash', '-c', 'tail -f /var/log/dmesg' ], 
   })
+  await container.rename({ 'name': containerNames.get('rename') })
+  const containerStatus = await container.status()
+  t.is(containerStatus.Name, '/' + containerNames.get('rename'))
+})
 
-  describe("#archive", function() {
-    it("should get an archive inside the container", function() {
-      this.timeout(70000)
-      return docker.container.create({ 
-        Image: 'ubuntu', 
-        Cmd: [ '/bin/bash', '-c', 'tail -f /var/log/dmesg' ], 
-        name: containerNames.get('get_archive') 
-      })
-        .then((container) => {
-          return container.fs.get({ path: '/var/log/dmesg' })
-        })
-        .then((data) => {
-          data.should.be.ok
-        })
-    });
-
-    it("should put an archive inside the container", function() {
-      this.timeout(70000)
-      return docker.container.create({ 
-        Image: 'ubuntu', 
-        Cmd: [ '/bin/bash', '-c', 'tail -f /var/log/dmesg' ], 
-        name: containerNames.get('put_archive') 
-      })
-        .then((container) => {
-          return container.fs.put('./test/test.tar', {
-            path: '/root'
-          })
-        })
-        .then((data) => {
-          data.should.be.ok
-        })
-    });
-
-    it("should inspect an archive inside the container", function() {
-      this.timeout(70000)
-      return docker.container.create({ 
-        Image: 'ubuntu', 
-        Cmd: [ '/bin/bash', '-c', 'tail -f /var/log/dmesg' ], 
-        name: containerNames.get('info_archive') 
-      })
-        .then((container) => {
-          return container.fs.info({ path: '/var/log/dmesg' })
-        })
-        .then((data) => {
-          data.should.be.ok
-        })
-    });
+test('pause', async t => {
+  const container = await createContainer('pause', {
+    Cmd: [ '/bin/bash', '-c', 'tail -f /var/log/dmesg' ], 
   })
+  await container.start()
+  await container.pause()
+  const containerStatus = await container.status()
+  t.is(containerStatus.State.Status, 'paused')
+  await container.unpause()
+  const containerStatus2 = await container.status()
+  t.is(containerStatus2.State.Status, 'running')
+})
 
-  after(function () {
-    this.timeout(30000)
-    var promises = []
-    containerNames.forEach(function (name) {
-      promises.push(
-        docker.container.stop(name)
-          .then((container) => {
-            return container.delete({ force: true })
-          })
-      )
-    })
-    return Promise.all(promises)
-      .catch((err) => console.log(err))
+test('commit', async t => {
+  const container = await createContainer('commit', {
+    Cmd: [ '/bin/bash', '-c', 'tail -f /var/log/dmesg' ], 
   })
+  await container.start()
+  const image = await container.commit({ comment: 'commit test' })
+  t.is(image.constructor, Image)
+})
+
+test('exec', async t => {
+  const container = await createContainer('exec', {
+    Cmd: [ '/bin/bash', '-c', 'tail -f /var/log/dmesg' ], 
+  })
+  await container.start()
+  const exec = await container.exec.create({
+    Cmd: [ "top" ]
+  })
+  const stream = await exec.start()
+  t.truthy(stream.pipe)
+})
+
+test('exec-status', async t => {
+  const container = await createContainer('inspect_exec', {
+    Cmd: [ '/bin/bash', '-c', 'tail -f /var/log/dmesg' ], 
+  })
+  await container.start()
+  const exec = await container.exec.create({
+    Cmd: [ "top" ]
+  })
+  const execStatus = await exec.status()
+  t.is(exec.constructor, execStatus.constructor)
+})
+
+test('attach', async t => {
+  const container = await createContainer('attach', {
+    'AttachStdin': false,
+    'AttachStdout': true,
+    'AttachStderr': true,
+    'Tty': false,
+    'OpenStdin': false,
+    'StdinOnce': false,
+    'Env': null,
+    'Cmd': ['/bin/bash', '-c', 'uptime'],
+    'Dns': ['8.8.8.8', '8.8.4.4'],
+    'Image': 'ubuntu',
+  })
+  const result = await container.attach({
+    stream: true,
+    stdout: true,
+    stderr: true
+  })
+  const stream = result[0]
+  t.truthy(stream)
+
+  await container.start()
+  const code = await container.wait()
+  t.is(code.StatusCode.constructor, Number)
+})
+
+test('get-archive', async t => {
+  const container = await createContainer('get_archive', {
+    Cmd: [ '/bin/bash', '-c', 'tail -f /var/log/dmesg' ], 
+  })
+  const data = await container.fs.get({ path: '/var/log/dmesg' })
+  t.truthy(data)
+})
+
+test('put-archive', async t => {
+  const container = await createContainer('put_archive', {
+    Cmd: [ '/bin/bash', '-c', 'tail -f /var/log/dmesg' ], 
+  })
+  const data = container.fs.put('./test/test.tar', {
+    path: '/root'
+  })
+  t.truthy(data)
+})
+
+test('inspect-archive', async t => {
+  const container = await createContainer('info_archive', {
+    Cmd: [ '/bin/bash', '-c', 'tail -f /var/log/dmesg' ], 
+  })
+  const data = await container.fs.info({ path: '/var/log/dmesg' })
+  t.truthy(data)
+})
+
+test.after.always('cleanup', async t => {
+  var promises = []
+  containerNames.forEach(function (name) {
+    if (name === 'docker_api_test_rename_prev') return
+    promises.push(
+      docker.container.stop(name)
+        .then((container) => {
+          return container.delete({ force: true })
+        })
+    )
+  })
+  t.notThrows(Promise.all(promises).catch((err) => console.log(err)))
 })
