+++ test/docker.js
@@ -1,52 +1,32 @@
-import chai from 'chai'
-import * as test_utils from './utils'
+import test from 'ava'
+import fs from 'fs'
+import { Docker } from '../lib/docker'
 
-const should = chai.should()
+const socket = process.env.DOCKER_SOCKET || '/var/run/docker.sock'
+const isSocket = fs.existsSync(socket) ? fs.statSync(socket).isSocket() : false
+const docker = isSocket
+  ? new Docker()
+  : new Docker({ socketPath: socket })
 
-const docker = test_utils.init()
+test('ping', async t => {
+  t.is(await docker.ping(), "OK")
+})
 
-describe('#docker', function() {
-  describe('#ping', function() {
-    it('should ping server', function () {
-      this.timeout(30000)
+test('version', async t => {
+  const data = await docker.version()
+  t.truthy(data)
+  t.is(data.Version.constructor, String)
+})
 
-      return docker.ping()
-    })
-  })
-  describe('#version', function() {
-    it('should retrieve version from server', function () {
-      this.timeout(30000)
+test('info', async t => {
+  const data = await docker.info()
+  t.truthy(data)
+  t.is(data.ServerVersion.constructor, String)
+})
 
-      return docker.version()
-        .then((data) => {
-          data.Version.should.be.a("string")
-        })
-    })
-  })
-  describe('#info', function() {
-    it('should retrieve info from server', function () {
-      this.timeout(30000)
-
-      return docker.info()
-        .then((data) => {
-          data.ServerVersion.should.be.a("string")
-        })
-    })
-  })
-  describe('#auth', function () {
-    it('should fail to auth', function () {
-      this.timeout(30000)
-
-      return docker.auth({
-        username: "AgustinCB",
-        password:" AgustinIsAwesome"
-      })
-        .then((data) => {
-          should.not.exist(data)
-        })
-        .catch((err) => {
-          err.should.be.ok
-        })
-    })
-  })
+test('auth', async t => {
+  t.throws(docker.auth({
+    username: "AgustinCB",
+    password: "AgustinIsAwesome"
+  }))
 })
